requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.1.5
djangorestframework==3.15.2
flake8==7.1.1
gunicorn==23.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
